feat : Form 삭제 (Request로 대체) / Request에서 입력받는 것으로 코드 변경
