--- src/main/java/spring/library/controller/form/CheckoutForm.java
@@ -1,10 +1,0 @@
-package spring.library.controller.form;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CheckoutForm {
-    private Long memberId;
-}
