+++ src/main/java/spring/library/controller/request/CheckoutRequest.java
@@ -1,4 +1,4 @@
-package spring.library.dto.request;
+package spring.library.controller.request;
 
 import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
@@ -6,15 +6,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Builder
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
 public class CheckoutRequest {
-    private Long bookId;
     private Long memberId;
-    private String loanDate;
-    private String dueDate;
-    private int renewalCount;
-    private boolean isReturned;
 }
