+++ src/main/java/spring/library/controller/request/PurchaseRequest.java
@@ -1,16 +1,17 @@
-package spring.library.controller.form;
+package spring.library.controller.request;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import spring.library.domain.ProcessResult;
 
 @Getter
 @NoArgsConstructor
-public class BookForm {
+public class PurchaseRequest {
+    private Long memberId;
     private String title;
     private String author;
     private String publisher;
     private int publicationYear;
-    private String classification;
-    private String status;
-    private int amount;
 }
