+++ src/main/java/spring/library/controller/request/PurchaseUpdateRequest.java
@@ -1,0 +1,11 @@
+package spring.library.controller.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class PurchaseUpdateRequest {
+    private String dateOfProcess;
+    private String processResult;
+}
